Intrusion Detection System (IDS) project/requirements.txt
numpy==1.22.4
pandas==1.4.2
scikit-learn==1.1.1
matplotlib==3.5.2
seaborn==0.11.2
jupyterlab==3.4.2
scapy==2.4.5
pyshark==0.5.3
elasticsearch==7.17.3
kibana==7.17.3
logstash==7.17.3
pytest==7.1.2
python-dotenv==0.20.0
flask==2.1.2
tensorflow==2.9.1
keras==2.9.0
joblib==1.1.0
pyyaml==6.0
